+++ R/val_pipeline.R
@@ -82,6 +82,7 @@ val_pipeline <- function(
   }
   apply_verbose(verbose)
   configure_bioc_repositories_if_requested(quiet = TRUE)
+  configure_riskmetric_offline_if_requested(quiet = TRUE)
 
   # Route pull_config() at any depth to the user-supplied config, if any.
   old_cfg <- options()["val.pipeline.config_path"]