+++ R/val_prep_pipeline.R
@@ -84,6 +84,7 @@ val_prep_pipeline <- function(
   stopifnot(inherits(as.Date(val_date), c("Date", "POSIXt")))
   apply_verbose(verbose)
   configure_bioc_repositories_if_requested(quiet = TRUE)
+  configure_riskmetric_offline_if_requested(quiet = TRUE)
 
   # Route pull_config() at any depth to the user-supplied config, if any.
   old_cfg <- options()["val.pipeline.config_path"]